Extract hierarchy creation to the mapper
