+++ lib/ui/src/modules/ui/components/left_panel/index.js
@@ -17,6 +17,7 @@ const mainStyle = {
 
 const storyProps = [
   'stories',
+  'storiesHierarchy',
   'selectedKind',
   'leftPanelHierarchy',
   'selectedStory',
