+++ lib/ui/src/modules/ui/components/left_panel/stories.js
@@ -71,58 +71,6 @@ class Stories extends React.Component {
     if (onSelectStory) onSelectStory(selectedKind, story);
   }
 
-  fillHierarchy(namespaces, hierarchy, story) {
-    if (namespaces.length === 1) {
-      const namespace = namespaces[0];
-      const childItems = hierarchy.map.get(namespace) || [];
-
-      childItems.push(story);
-      hierarchy.map.set(namespace, childItems);
-      return;
-    }
-
-    const namespace = namespaces[0];
-    const childItems = hierarchy.map.get(namespace) || [];
-    let childHierarchy = childItems.find(item => item.isNamespace);
-
-    if (!childHierarchy) {
-      childHierarchy = {
-        isNamespace: true,
-        current: namespace,
-        namespace: hierarchy.namespace ? `${hierarchy.namespace}${namespace}.` : `${namespace}.`,
-        firstKind: story.kind,
-        map: new Map(),
-      };
-
-      childItems.push(childHierarchy);
-      hierarchy.map.set(namespace, childItems);
-    }
-
-    this.fillHierarchy(namespaces.slice(1), childHierarchy, story);
-  }
-
-  createStoriesHierarchy(stories) {
-    const groupedStories = stories.map(story => {
-      const namespaces = story.kind.split('.');
-
-      return {
-        namespaces,
-        name: namespaces[namespaces.length - 1],
-        ...story,
-      };
-    });
-
-    const hierarchyRoot = {
-      namespace: '',
-      current: '',
-      map: new Map(),
-    };
-
-    groupedStories.forEach(story => this.fillHierarchy(story.namespaces, hierarchyRoot, story));
-
-    return hierarchyRoot;
-  }
-
   renderMenuItem(item, style, onClick, displayName) {
     return (
       <a title={`Open ${item}`} style={style} onClick={onClick} role="menuitem" tabIndex="0">
@@ -217,14 +165,12 @@ class Stories extends React.Component {
   }
 
   render() {
-    const { stories, leftPanelHierarchy } = this.props;
+    const { stories, storiesHierarchy, leftPanelHierarchy } = this.props;
 
-    if (leftPanelHierarchy) {
-      const hierarchy = this.createStoriesHierarchy(stories);
-
+    if (leftPanelHierarchy && storiesHierarchy) {
       return (
         <div style={listStyle}>
-          {this.renderHierarchy(hierarchy)}
+          {this.renderHierarchy(storiesHierarchy)}
         </div>
       );
     }
@@ -241,7 +187,9 @@ class Stories extends React.Component {
 
 Stories.defaultProps = {
   stories: [],
+  leftPanelHierarchy: false,
   onSelectStory: null,
+  storiesHierarchy: null,
 };
 
 Stories.propTypes = {
@@ -251,6 +199,11 @@ Stories.propTypes = {
       stories: PropTypes.array,
     })
   ),
+  storiesHierarchy: PropTypes.shape({
+    namespace: PropTypes.string,
+    current: PropTypes.string,
+    map: PropTypes.object,
+  }),
   selectedKind: PropTypes.string.isRequired,
   selectedStory: PropTypes.string.isRequired,
   leftPanelHierarchy: PropTypes.bool.isRequired,
